refactor: カスタムフックからfetchを切り出し
