+++ app/client/deletePost.ts
@@ -1,0 +1,10 @@
+export const deletePost = async (contentId: string) => {
+  const res = await fetch(`${process.env.NEXT_PUBLIC_BASE_URL}/api/posts`, {
+    method: 'DELETE',
+    headers: {
+      'Content-Type': 'application/json',
+    },
+    body: JSON.stringify({ contentId }),
+  })
+  if (res.status !== 200) throw new Error('削除に失敗しました')
+}
