+++ app/client/loginFirebase.ts
@@ -1,0 +1,11 @@
+export const loginFirebase = async (email: string, password: string) => {
+  const res = await fetch(`${process.env.NEXT_PUBLIC_BASE_URL}/api/login`, {
+    method: 'POST',
+    headers: {
+      'Content-Type': 'application/json',
+    },
+    body: JSON.stringify({ email, password }),
+  })
+  const data = await res.json()
+  return data
+}
