+++ app/client/useCreatePost.ts
@@ -1,16 +1,4 @@
 import { useMutation } from '@tanstack/react-query'
+import { createPost } from './createPost'
 
 export const useCreatePost = () => useMutation((question: string) => createPost(question))
-
-export const createPost = async (question: string) => {
-  const res = await fetch(`${process.env.NEXT_PUBLIC_BASE_URL}/api/posts`, {
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/json',
-    },
-    body: JSON.stringify({ question }),
-  })
-  const data: { id: number; error: string } = await res.json()
-  if (res.status !== 200) throw new Error(data.error)
-  return data
-}
