+++ app/client/useMoveToAnswer.ts
@@ -1,6 +1,6 @@
 import { useMutation } from '@tanstack/react-query'
-import { createPost } from './useCreatePost'
-import { deleteReply } from './useDeleteReply'
+import { deleteReply } from './deleteReply'
+import { createPost } from './createPost'
 
 export const useMoveToAnswer = () =>
   useMutation(async ({ contentId, question }: { contentId: string; question: string }) => {
